WIP handle grapheme seams.
Because the text is stored in a Rope split up between leaf nodes, it's
possible for a grapheme to be split across leaf boundaries.  When this
happens, the Rope will no longer recognize it as a single grapheme.

This is working towards resolving that issue.  The idea is to check for
and repair such seams in the operations that can produce them.